allow to delete owner only wheter he has no animals registered

+++ services/owner.js
@@ -1,4 +1,5 @@
 import ownerRepository from "../repositories/owner.js";
+import animalRepository from "../repositories/animal.js";
 
 async function insertOwner (owner) {
   return ownerRepository.insertOwner(owner)
@@ -19,6 +20,10 @@ async function updateOwner (owner) {
 async function deleteOwner (id) {
   const owner = await getOwnerById(id);
   if (!owner) throw new Error('The Owner with the given ID was not found.');
+
+  const animalsByOwner =  await animalRepository.getAnimalByOwner(id)
+  if (animalsByOwner.length > 0)  throw new Error('The Owner with the given ID has animals registered.');
+
   return await ownerRepository.deleteOwner(id)
 }
